implemented TAP output
Writes various test result data as YAML, so there's some careful
indentation going on.

I tried using pprint-logical-block, but that has side effects all over
the place and was affecting printing even outside the logical block,
so I'm not sure how that's intended to be used.  There's a format
string equivalent that's messier but at least it works.